Add "Aim" to ranged weapon dialogs
