+++ style/dialog/dialog.css
@@ -130,7 +130,7 @@
 
 
 .wrath-and-glory.dialog .wrapper .range {
-  flex: 0.73
+  flex: 0.5
 }
 
 .wrath-and-glory.dialog .wrapper .die {
