+++ template/item/armour.html
@@ -67,7 +67,7 @@
           <div class="item-form-group static">
             <label>{{localize "ARMOUR.INVULNERABLE"}}</label>
             <div class="item-form-inputs">
-              <div class="item-checkbox {{#if data.invulnerable}}checked{{else}}unchecked{{/if}}" data-target="data.invulnerable">
+              <div class="wng-checkbox {{#if data.invulnerable}}checked{{else}}unchecked{{/if}}" data-target="data.invulnerable">
               {{#if data.invulnerable}}
                 <i class="fas fa-check"></i>
               {{/if}}
