+++ template/item/psychicPower.html
@@ -77,7 +77,7 @@
           <div class="item-form-group static">
             <label>{{localize "PSYCHIC_POWER.MULTI_TARGET"}}</label>
             <div class="item-form-inputs">
-              <div class="item-checkbox {{#if data.multiTarget}}checked{{else}}unchecked{{/if}}" data-target="data.multiTarget">
+              <div class="wng-checkbox {{#if data.multiTarget}}checked{{else}}unchecked{{/if}}" data-target="data.multiTarget">
               {{#if data.multiTarget}}
                 <i class="fas fa-check"></i>
               {{/if}}
@@ -233,7 +233,7 @@
                               
                               <input type="number" step = "any" data-path="value" placeholder='{{localize "POWER.POTENCY_VALUE"}}' value="{{potency.value}}">
                               
-                              <div class="item-checkbox {{#if potency.single}}checked{{else}}unchecked{{/if}}" data-path="single" data-target="potency">
+                              <div class="wng-checkbox {{#if potency.single}}checked{{else}}unchecked{{/if}}" data-path="single" data-target="potency">
                                 {{#if potency.single}}
                                   <i class="fas fa-check"></i>
                                 {{/if}}
