+++ template/item/talent.html
@@ -47,7 +47,7 @@
           <div class="item-form-group static">
             <label>{{localize "TALENT.DISPLAY"}}</label>
             <div class="item-form-inputs">
-              <div class="item-checkbox {{#if data.display}}checked{{else}}unchecked{{/if}}" data-target="data.display">
+              <div class="wng-checkbox {{#if data.display}}checked{{else}}unchecked{{/if}}" data-target="data.display">
               {{#if data.display}}
                 <i class="fas fa-check"></i>
               {{/if}}
